# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AfriLance freelance marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles/me:
    get:
      operationId: getMyProfile
      tags: [profiles]
      summary: Get current user's profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Profile not found
    put:
      operationId: upsertMyProfile
      tags: [profiles]
      summary: Create or update current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profiles/{id}:
    get:
      operationId: getProfile
      tags: [profiles]
      summary: Get a profile by user ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Not found
  /profiles/freelancers:
    get:
      operationId: listFreelancers
      tags: [profiles]
      summary: List freelancer profiles with filters
      parameters:
        - name: skill
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: minRate
          in: query
          schema:
            type: number
        - name: maxRate
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of freelancers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreelancerListResponse"

  # Jobs
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payment transactions for current user
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentListResponse"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Initialize a Paystack payment and create escrow record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Payment initialized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InitializePaymentResponse"
  /payments/verify/{reference}:
    get:
      operationId: verifyPayment
      tags: [payments]
      summary: Verify a Paystack payment and mark as escrowed
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}/release:
    patch:
      operationId: releasePayment
      tags: [payments]
      summary: Release escrowed funds to freelancer via Paystack transfer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReleasePaymentBody"
      responses:
        "200":
          description: Payment released
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
components:
  schemas:
...
    Profile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        role:
          type: string
        name:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        hourlyRate:
          type: ["number", "null"]
        category:
          type: ["string", "null"]
        portfolioItems:
          type: array
          items:
            $ref: "#/components/schemas/PortfolioItem"
        avgRating:
          type: ["number", "null"]
        totalReviews:
          type: integer
        featured:
          type: boolean
        isBlocked:
          type: boolean
        isVerified:
          type: boolean
        isTopRated:
          type: boolean
        resumeUrl:
          type: ["string", "null"]
        verificationDocUrl:
          type: ["string", "null"]
        verificationStatus:
          type: string
        completedJobs:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - userId
        - role
        - name
        - skills
        - portfolioItems
        - totalReviews
        - featured
        - isBlocked
        - isVerified
        - isTopRated
        - verificationStatus
        - completedJobs
        - createdAt
    PortfolioItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        projectUrl:
          type: ["string", "null"]
      required:
        - id
        - title
    UpsertProfileBody:
      type: object
      properties:
        role:
          type: string
        name:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        hourlyRate:
          type: ["number", "null"]
        category:
          type: ["string", "null"]
        portfolioItems:
          type: array
          items:
            $ref: "#/components/schemas/PortfolioItem"
        resumeUrl:
          type: ["string", "null"]
        verificationDocUrl:
          type: ["string", "null"]
      required:
        - role
        - name
    FreelancerListResponse:
      type: object
      properties:
        freelancers:
          type: array
          items:
            $ref: "#/components/schemas/Profile"
        total:
          type: integer
      required:
        - freelancers
        - total
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        clientId:
          type: integer
        freelancerId:
          type: integer
        amount:
          type: number
        status:
          type: string
        paystackReference:
          type: ["string", "null"]
        paystackTransferCode:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        freelancerName:
          type: ["string", "null"]
        clientName:
          type: ["string", "null"]
        createdAt:
          type: string
        releasedAt:
          type: ["string", "null"]
      required:
        - id
        - jobId
        - clientId
        - freelancerId
        - amount
        - status
        - createdAt
    CreatePaymentBody:
      type: object
      properties:
        jobId:
          type: integer
        freelancerId:
          type: integer
        amount:
          type: number
        clientEmail:
          type: string
      required:
        - jobId
        - freelancerId
        - amount
        - clientEmail
    InitializePaymentResponse:
      type: object
      properties:
        paymentId:
          type: integer
        authorizationUrl:
          type: string
        accessCode:
          type: string
        reference:
          type: string
      required:
        - paymentId
        - authorizationUrl
        - accessCode
        - reference
    ReleasePaymentBody:
      type: object
      properties:
        recipientCode:
          type: string
        accountNumber:
          type: string
        bankCode:
          type: string
        accountName:
          type: string
    PaymentListResponse:
      type: object
      properties:
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        total:
          type: integer
      required:
        - payments
        - total
...