# lib/api-zod/src/generated/types/payment.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Api
 * AfriLance freelance marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Payment {
  id: number;
  jobId: number;
  clientId: number;
  freelancerId: number;
  amount: number;
  status: string;
  /** @nullable */
  paystackReference?: string | null;
  /** @nullable */
  paystackTransferCode?: string | null;
  /** @nullable */
  jobTitle?: string | null;
  /** @nullable */
  freelancerName?: string | null;
  /** @nullable */
  clientName?: string | null;
  createdAt: string;
  /** @nullable */
  releasedAt?: string | null;
}